@@ -1,24 +1,28 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.janusgraph</groupId>
     <artifactId>janusgraph-foundationdb</artifactId>
-    <version>0.1.0</version>
+    <version>0.2.0-SNAPSHOT</version>
     <packaging>jar</packaging>
     <name>JanusGraph FoundationDB Storage Adapter</name>
 
     <properties>
-        <maven.compiler.plugin.version>3.8.1</maven.compiler.plugin.version>
-        <jdk.version>1.8</jdk.version>
-        <janusgraph.version>0.5.2</janusgraph.version>
-        <foundationdb.version>6.2.22</foundationdb.version>
-        <dependency.plugin.version>3.2.0</dependency.plugin.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <maven.compiler.plugin.version>3.11.0</maven.compiler.plugin.version>
+        <jdk.version>21</jdk.version>
+        <janusgraph.version>1.1.0</janusgraph.version>
+        <foundationdb.version>7.4.1</foundationdb.version>
+        <slf4j.version>2.0.17</slf4j.version>
+        <dependency.plugin.version>3.6.0</dependency.plugin.version>
+        <surefire.plugin.version>3.1.2</surefire.plugin.version>
+        <assembly.plugin.version>3.6.0</assembly.plugin.version>
         <test.skip.default>false</test.skip.default>
         <test.skip.tp>true</test.skip.tp>
-        <testcontainers.version>1.15.3</testcontainers.version>
+        <testcontainers.version>1.21.0</testcontainers.version>
     </properties>
 
     <developers>
