@@ -1,4 +1,4 @@
-// Copyright 2018 Expero Inc.
+// Copyright 2018 Expero Inc., 2023 JanusGraph Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,9 +15,7 @@
 package org.janusgraph.diskstorage.foundationdb;
 
 import com.apple.foundationdb.KeyValue;
-import com.apple.foundationdb.LocalityUtil;
 import com.apple.foundationdb.async.AsyncIterator;
-import com.apple.foundationdb.async.CloseableAsyncIterator;
 import com.apple.foundationdb.directory.DirectorySubspace;
 import com.google.common.base.Preconditions;
 import org.janusgraph.diskstorage.BackendException;
@@ -32,35 +30,36 @@ import org.janusgraph.diskstorage.util.StaticArrayBuffer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.HashMap;
+import java.util.stream.Collectors;
 
 /**
+ * FoundationDB implementation of {@link OrderedKeyValueStore}.
+ * Delegates operations to {@link FoundationDBTx}.
+ *
  * @author Ted Wilmes (twilmes@gmail.com)
+ * @author JanusGraph Authors
  */
 public class FoundationDBKeyValueStore implements OrderedKeyValueStore, AutoCloseable {
 
     private static final Logger log = LoggerFactory.getLogger(FoundationDBKeyValueStore.class);
 
-    static final StaticBuffer.Factory<byte[]> ENTRY_FACTORY = (array, offset, limit) -> {
-        final byte[] bArray = new byte[limit - offset];
-        System.arraycopy(array, offset, bArray, 0, limit - offset);
-        return bArray;
-    };
+    static final StaticBuffer.Factory<byte[]> ENTRY_FACTORY = Arrays::copyOfRange;
 
-    private final DirectorySubspace db;
+    private final DirectorySubspace dbSubspace; // Represents the specific FDB subspace for this store
     private final String name;
     private final FoundationDBStoreManager manager;
-    private boolean isOpen;
+    private volatile boolean isOpen;
 
-    public FoundationDBKeyValueStore(String n, DirectorySubspace data, FoundationDBStoreManager m) {
-        db = data;
-        name = n;
-        manager = m;
-        isOpen = true;
+    public FoundationDBKeyValueStore(String name, DirectorySubspace subspace, FoundationDBStoreManager manager) {
+        this.dbSubspace = Preconditions.checkNotNull(subspace);
+        this.name = Preconditions.checkNotNull(name);
+        this.manager = Preconditions.checkNotNull(manager);
+        this.isOpen = true;
+        log.debug("FoundationDBKeyValueStore '{}' initialized with subspace prefix: {}", name, subspace.getKey());
     }
 
     @Override
@@ -68,38 +67,35 @@ public class FoundationDBKeyValueStore implements OrderedKeyValueStore, AutoClos
         return name;
     }
 
-    private static FoundationDBTx getTransaction(StoreTransaction txh) {
-        Preconditions.checkArgument(txh != null);
-        return ((FoundationDBTx) txh);//.getTransaction();
-    }
-
     @Override
     public synchronized void close() throws BackendException {
-        if (isOpen) manager.removeDatabase(this);
-        isOpen = false;
+        // Use synchronized for simple state check/update, low contention expected here.
+        if (isOpen) {
+            log.info("Closing FoundationDBKeyValueStore: {}", name);
+            isOpen = false;
+            manager.removeDatabase(this); // Notify manager
+            // dbSubspace doesn't need explicit closing, FDB connection is managed by StoreManager
+        } else {
+            log.warn("FoundationDBKeyValueStore '{}' already closed.", name);
+        }
+    }
+
+    private void ensureOpen() throws PermanentBackendException {
+        if (!isOpen) {
+            throw new PermanentBackendException("Store '" + name + "' has been closed.");
+        }
     }
 
     @Override
     public StaticBuffer get(StaticBuffer key, StoreTransaction txh) throws BackendException {
-        FoundationDBTx tx = getTransaction(txh);
-        try {
-            byte[] databaseKey = db.pack(key.as(ENTRY_FACTORY));
-            log.trace("db={}, op=get, tx={}", name, txh);
-            final byte[] entry = tx.get(databaseKey);
-            if (entry != null) {
-                return getBuffer(entry);
-            } else {
-                return null;
-            }
-        }
-        catch (BackendException e) {
-            log.error ("db={}, op=get, tx={} with exception", name, txh, e);
-            throw e;
-        }
-        catch (Exception e) {
-            log.error ("db={}, op=get, tx={} with exception", name, txh, e);
-            throw new PermanentBackendException(e);
-        }
+        ensureOpen();
+        FoundationDBTx tx = FoundationDBTx.getTx(txh);
+        byte[] databaseKey = dbSubspace.pack(key.as(ENTRY_FACTORY));
+        log.trace("Store '{}', op=get, key={}, tx={}", name, key, tx);
+
+        byte[] valueBytes = tx.get(databaseKey);
+
+        return (valueBytes != null) ? getBuffer(valueBytes) : null;
     }
 
     @Override
@@ -109,170 +105,163 @@ public class FoundationDBKeyValueStore implements OrderedKeyValueStore, AutoClos
 
     @Override
     public void acquireLock(StaticBuffer key, StaticBuffer expectedValue, StoreTransaction txh) throws BackendException {
-        if (getTransaction(txh) == null) {
-            log.warn("Attempt to acquire lock with transactions disabled");
-        } //else we need no locking
-    }
-
-    @Override
-    public void insert(StaticBuffer key, StaticBuffer value, StoreTransaction txh, Integer integer) throws BackendException {
-        insert(key, value, txh);
+        // FoundationDB uses Optimistic Concurrency Control with transaction retries.
+        // JanusGraph's explicit locking via acquireLock is generally NOT needed or recommended
+        // when using FDB's serializable isolation. Conflicts are detected at commit time.
+        // If using weaker isolation, application-level locking might be needed, but that
+        // pattern is discouraged with FDB.
+        // For compatibility, we don't throw, but log a warning.
+        if (FoundationDBTx.getTx(txh).getIsolationLevel() == FoundationDBTx.IsolationLevel.SERIALIZABLE) {
+            log.debug("Store '{}', op=acquireLock (ignored - using FDB optimistic locking), key={}, tx={}", name, key, txh);
+        } else {
+            log.warn("Store '{}', op=acquireLock called with non-serializable isolation level. " +
+                "FDB relies on transaction conflicts, explicit locking may be ineffective or lead to deadlocks. Key={}, tx={}", name, key, txh);
+        }
+        // No operation needed with FDB's model.
     }
 
     @Override
     public RecordIterator<KeyValueEntry> getSlice(KVQuery query, StoreTransaction txh) throws BackendException {
-        if (manager.getMode() == FoundationDBStoreManager.RangeQueryIteratorMode.SYNC) {
-            return getSliceSync(query, txh);
+        ensureOpen();
+        FoundationDBTx tx = FoundationDBTx.getTx(txh);
+        var rangeQuery = new FoundationDBRangeQuery(dbSubspace, query);
+
+        // Delegate based on configured mode in the manager
+        if (manager.getRangeQueryMode() == FoundationDBStoreManager.RangeQueryIteratorMode.ASYNC) {
+            return getSliceAsync(rangeQuery, tx);
         } else {
-            return getSliceAsync(query, txh);
+            return getSliceSync(rangeQuery, tx);
         }
     }
 
-    public RecordIterator<KeyValueEntry> getSliceSync(KVQuery query, StoreTransaction txh) throws BackendException {
-        log.trace("beginning db={}, op=getSliceSync, tx={}", name, txh);
+    private RecordIterator<KeyValueEntry> getSliceSync(FoundationDBRangeQuery rangeQuery, FoundationDBTx tx) throws BackendException {
+        KVQuery query = rangeQuery.asKVQuery();
+        log.trace("Store '{}', op=getSliceSync, query={}, tx={}", name, query, tx);
 
-        final FoundationDBTx tx = getTransaction(txh);
+        List<KeyValue> resultList = tx.getRange(rangeQuery); // Uses retry logic internally
 
-        try {
-            final List<KeyValue> result = tx.getRange(new FoundationDBRangeQuery(db, query));
+        log.debug("Store '{}', op=getSliceSync, query={}, tx={}, resultCount={}", name, query, tx, resultList.size());
 
-            log.trace("db={}, op=getSliceSync, tx={}, result-count={}", name, txh, result.size());
-
-            return new FoundationDBRecordIterator(db, result.iterator(), query.getKeySelector());
-
-        }
-        catch (BackendException e) {
-            throw e;
-        }
-        catch (Exception e) {
-            log.error("db={}, op=getSliceSync, tx={} with exception", name, txh, e);
-            throw new PermanentBackendException(e);
-        }
+        // Wrap the in-memory list with a standard iterator
+        return new FoundationDBRecordIterator(dbSubspace, resultList.iterator(), query.getKeySelector());
     }
 
-
-    public RecordIterator<KeyValueEntry> getSliceAsync(KVQuery query, StoreTransaction txh) throws BackendException {
-
-        log.trace("db={}, op=getSliceAsync, tx={}", name, txh);
-
-        final FoundationDBTx tx = getTransaction(txh);
+    private RecordIterator<KeyValueEntry> getSliceAsync(FoundationDBRangeQuery rangeQuery, FoundationDBTx tx) throws BackendException {
+        KVQuery query = rangeQuery.asKVQuery();
+        log.trace("Store '{}', op=getSliceAsync, query={}, tx={}", name, query, tx);
 
         try {
-            final FoundationDBRangeQuery rangeQuery = new FoundationDBRangeQuery(db, query);
-            final AsyncIterator<KeyValue> result = tx.getRangeIter( rangeQuery);
+            // Get the FDB async iterator (does not use retry logic itself)
+            AsyncIterator<KeyValue> fdbIterator = tx.getRangeIterator(rangeQuery);
 
-            return new FoundationDBRecordAsyncIterator(db, tx, rangeQuery, result, query.getKeySelector());
+            // Wrap FDB's async iterator
+            // Pass the *transaction* context in case it's needed (though removed retry logic from iterator itself)
+            return new FoundationDBRecordAsyncIterator(dbSubspace, tx, rangeQuery, fdbIterator, query.getKeySelector());
+        } catch (BackendException e) {
+            log.error("Store '{}', op=getSliceAsync failed during iterator setup. Query={}, tx={}", name, query, tx, e);
+            throw e; // Re-throw known BackendExceptions
         } catch (Exception e) {
-            log.error("getSliceAsync db={}, tx={} with exception", name, txh, e);
-            throw new PermanentBackendException(e);
+            log.error("Store '{}', op=getSliceAsync failed with unexpected error during iterator setup. Query={}, tx={}", name, query, tx, e);
+            throw new PermanentBackendException("Failed to initialize async slice iterator", e);
         }
     }
-
 
     @Override
-    public Map<KVQuery,RecordIterator<KeyValueEntry>> getSlices(List<KVQuery> queries, StoreTransaction txh) throws BackendException {
-        if (manager.getMode() == FoundationDBStoreManager.RangeQueryIteratorMode.SYNC) {
-            return getSlicesSync(queries, txh);
+    public Map<KVQuery, RecordIterator<KeyValueEntry>> getSlices(List<KVQuery> queries, StoreTransaction txh) throws BackendException {
+        ensureOpen();
+        FoundationDBTx tx = FoundationDBTx.getTx(txh);
+
+        if (manager.getRangeQueryMode() == FoundationDBStoreManager.RangeQueryIteratorMode.ASYNC) {
+            return getSlicesAsync(queries, tx);
         } else {
-            return getSlicesAsync(queries, txh);
-        }
-    }
-
-    public Map<KVQuery,RecordIterator<KeyValueEntry>> getSlicesSync (List<KVQuery> queries, StoreTransaction txh) throws BackendException {
-        log.trace("beginning db={}, op=getSlicesSync, tx={}", name, txh);
-        FoundationDBTx tx = getTransaction(txh);
-        final Map<KVQuery, FoundationDBRangeQuery> fdbQueries = new HashMap<>();
-
-        try {
-            for (final KVQuery q : queries) {
-                fdbQueries.put(q, new FoundationDBRangeQuery(db, q));
-            }
-
-            final Map<KVQuery, List<KeyValue>> unfilteredResultMap = tx.getMultiRange(fdbQueries.values());
-            final Map<KVQuery, RecordIterator<KeyValueEntry>> iteratorMap = new HashMap<>();
-            for (Map.Entry<KVQuery, List<KeyValue>> kv : unfilteredResultMap.entrySet()) {
-                iteratorMap.put(kv.getKey(),
-                        new FoundationDBRecordIterator(db, kv.getValue().iterator(), kv.getKey().getKeySelector()));
-            }
-
-            return iteratorMap;
-        }
-        catch (BackendException e){
-            throw e;
-        }
-        catch (Exception e) {
-            log.error("db={}, op=getSlicesSync, tx={} throws exception", name, txh, e);
-            throw new PermanentBackendException(e);
+            return getSlicesSync(queries, tx);
         }
     }
 
 
-    public Map<KVQuery,RecordIterator<KeyValueEntry>> getSlicesAsync(List<KVQuery> queries, StoreTransaction txh) throws BackendException {
-        log.trace("beginning db={}, op=getSlicesAsync, tx={}", name, txh);
+    private Map<KVQuery, RecordIterator<KeyValueEntry>> getSlicesSync(List<KVQuery> queries, FoundationDBTx tx) throws BackendException {
+        log.trace("Store '{}', op=getSlicesSync, queries={}, tx={}", name, queries.size(), tx);
+        Map<KVQuery, FoundationDBRangeQuery> fdbQueries = queries.stream()
+            .collect(Collectors.toMap(q -> q, q -> new FoundationDBRangeQuery(dbSubspace, q)));
 
-        FoundationDBTx tx = getTransaction(txh);
-        final Map<KVQuery, RecordIterator<KeyValueEntry>> resultMap = new HashMap<>();
+        Map<KVQuery, List<KeyValue>> resultMap = tx.getMultiRange(fdbQueries.values()); // Uses retry logic
+
+        Map<KVQuery, RecordIterator<KeyValueEntry>> iteratorMap = new HashMap<>(resultMap.size());
+        resultMap.forEach((query, kvList) -> {
+            iteratorMap.put(query, new FoundationDBRecordIterator(dbSubspace, kvList.iterator(), query.getKeySelector()));
+        });
+
+        log.debug("Store '{}', op=getSlicesSync, queries={}, tx={}, resultSets={}", name, queries.size(), tx, iteratorMap.size());
+        return iteratorMap;
+    }
+
+    private Map<KVQuery, RecordIterator<KeyValueEntry>> getSlicesAsync(List<KVQuery> queries, FoundationDBTx tx) throws BackendException {
+        log.trace("Store '{}', op=getSlicesAsync, queries={}, tx={}", name, queries.size(), tx);
+        Map<KVQuery, RecordIterator<KeyValueEntry>> resultMap = new HashMap<>(queries.size());
 
         try {
             for (final KVQuery query : queries) {
-                FoundationDBRangeQuery rangeQuery = new FoundationDBRangeQuery(db, query);
-                AsyncIterator<KeyValue> result = tx.getRangeIter(rangeQuery);
-                resultMap.put(query,
-                        new FoundationDBRecordAsyncIterator(db, tx, rangeQuery, result, query.getKeySelector()));
+                FoundationDBRangeQuery rangeQuery = new FoundationDBRangeQuery(dbSubspace, query);
+                // Get iterator, handle exceptions during setup
+                AsyncIterator<KeyValue> fdbIterator = tx.getRangeIterator(rangeQuery);
+                resultMap.put(query, new FoundationDBRecordAsyncIterator(dbSubspace, tx, rangeQuery, fdbIterator, query.getKeySelector()));
             }
+        } catch (BackendException e) {
+            log.error("Store '{}', op=getSlicesAsync failed during iterator setup for one or more queries. Tx={}", name, tx, e);
+            // Clean up any iterators already created? Difficult. Re-throw.
+            resultMap.values().forEach(iter -> { try { iter.close(); } catch (Exception ignored) {} });
+            throw e;
         } catch (Exception e) {
-            log.error("db={}, getSlicesAsync, tx= {} with exception", name, txh, e);
-
-            throw new PermanentBackendException(e);
+            log.error("Store '{}', op=getSlicesAsync failed with unexpected error during setup. Tx={}", name, tx, e);
+            resultMap.values().forEach(iter -> { try { iter.close(); } catch (Exception ignored) {} });
+            throw new PermanentBackendException("Failed to initialize one or more async slice iterators", e);
         }
 
-
+        log.debug("Store '{}', op=getSlicesAsync, queries={}, tx={}, iteratorsCreated={}", name, queries.size(), tx, resultMap.size());
         return resultMap;
     }
 
-    public void insert(StaticBuffer key, StaticBuffer value, StoreTransaction txh) throws BackendException {
-        FoundationDBTx tx = getTransaction(txh);
-        try {
-
-            log.trace("db={}, op=insert, tx={}", name, txh);
-            tx.set(db.pack(key.as(ENTRY_FACTORY)), value.as(ENTRY_FACTORY));
-        } catch (Exception e) {
-            log.error ("db={}, op=insert, tx={} with exception", name, txh, e);
-            throw new PermanentBackendException(e);
-        }
-    }
 
     @Override
+    public void insert(StaticBuffer key, StaticBuffer value, StoreTransaction txh, Integer ttl) throws BackendException {
+        // FoundationDB doesn't directly support TTLs at the storage layer.
+        // TTLs in JanusGraph are typically handled by background jobs checking timestamps.
+        if (ttl != null && ttl > 0) {
+            log.warn("Store '{}', op=insert with TTL={} requested, but FoundationDB backend does not natively support TTL. TTL ignored. Key={}", name, ttl, key);
+        }
+        insert(key, value, txh); // Call the non-TTL version
+    }
+
+    // Internal insert implementation called by mutateMany and public insert variants
+    void insert(StaticBuffer key, StaticBuffer value, FoundationDBTx tx) throws BackendException {
+        ensureOpen();
+        byte[] dbKey = dbSubspace.pack(key.as(ENTRY_FACTORY));
+        byte[] dbValue = value.as(ENTRY_FACTORY);
+        log.trace("Store '{}', op=insert, key={}, valueSize={}, tx={}", name, key, dbValue.length, tx);
+        tx.set(dbKey, dbValue);
+    }
+
+    // Public interface version
+    public void insert(StaticBuffer key, StaticBuffer value, StoreTransaction txh) throws BackendException {
+        insert(key, value, FoundationDBTx.getTx(txh));
+    }
+
+    // Internal delete implementation called by mutateMany
+    void delete(StaticBuffer key, FoundationDBTx tx) throws BackendException {
+        ensureOpen();
+        byte[] dbKey = dbSubspace.pack(key.as(ENTRY_FACTORY));
+        log.trace("Store '{}', op=delete, key={}, tx={}", name, key, tx);
+        tx.clear(dbKey);
+    }
+
+    // Public interface version
+    @Override
     public void delete(StaticBuffer key, StoreTransaction txh) throws BackendException {
-        FoundationDBTx tx = getTransaction(txh);
-        try {
-            log.trace("db={}, op=delete, tx={}", name, txh);
-            tx.clear(db.pack(key.as(ENTRY_FACTORY)));
-        } catch (Exception e) {
-            log.error("db={}, op=delete, tx={} with exception", name, txh, e);
-            throw new PermanentBackendException(e);
-        }
+        delete(key, FoundationDBTx.getTx(txh));
     }
 
-    /* this method is not called within JanusGraph, it's used for bulk access in Spark jobs.
-    For more: Please check: {@link https://github.com/JanusGraph/janusgraph-foundationdb/pull/48#discussion_r498917927}
-     */
-    public List<StaticBuffer> getBoundaryKeys() {
-        List<StaticBuffer> keys = new ArrayList<>();
-        try (CloseableAsyncIterator<byte[]> it = LocalityUtil.getBoundaryKeys(manager.db, db.range().begin, db.range().end)) {
-            it.forEachRemaining(key -> {
-                if (key[key.length - 1] != 0x00) {
-                    key = Arrays.copyOf(key, key.length + 1);
-                }
-                keys.add(getBuffer(db.unpack(key).getBytes(0)));
-            });
-        }
-        return keys;
-    }
-
+    // Helper to convert byte[] to StaticBuffer using efficient buffer pooling (StaticArrayBuffer does this)
     static StaticBuffer getBuffer(byte[] entry) {
-        return new StaticArrayBuffer(entry);
+        return StaticArrayBuffer.of(entry);
     }
-
-
 }
